backend/pyproject.toml
[tool.poetry]
name = "trader-helper-backend"
version = "0.1.0"
description = "Trader Helper backend (FastAPI + Celery)"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {version = "^0.27.0", extras = ["standard"]}
celery = {version = "^5.3.6", extras = ["redis"]}
redis = "^5.0.4"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.25"
psycopg = {version = "^3.1.18", extras = ["binary"]}
httpx = "^0.27.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.0"

[tool.poetry.scripts]
start-api = "uvicorn app.main:app --host 0.0.0.0 --port 8000"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"
